backend/requirements/development.txt
-r base.txt

# Development tools
django-debug-toolbar>=4.2.0
django-extensions>=3.2.3
ipython>=8.18.0

# Testing
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Code quality
black>=23.12.0
flake8>=6.1.0
isort>=5.13.0
